+++ e2e/commands/export.e2e.1.js
@@ -831,6 +831,7 @@ describe('bit export command', function () {
             helper.scopeHelper.reInitRemoteScope(forkScopePath);
             helper.fixtures.createComponentUtilsIsString(fixtures.isStringModulePath(helper.scopes.remote));
             helper.fixtures.createComponentBarFoo(fixtures.barFooModulePath(helper.scopes.remote));
+            helper.env.importDummyCompiler();
             helper.command.tagScope('1.0.0');
             helper.command.export(`${forkScope} --include-dependencies --force --codemod`);
           });
@@ -856,6 +857,32 @@ describe('bit export command', function () {
               expect(barFoo).to.have.string(forkScope);
               expect(barFoo).to.not.have.string(helper.scopes.remote);
             });
+          });
+        });
+        describe('with --set-current-scope', () => {
+          before(() => {
+            helper.scopeHelper.getClonedLocalScope(localScope);
+            helper.scopeHelper.reInitRemoteScope(forkScopePath);
+            helper.fixtures.createComponentUtilsIsString(fixtures.isStringModulePath(helper.scopes.remote));
+            helper.fixtures.createComponentBarFoo(fixtures.barFooModulePath(helper.scopes.remote));
+            helper.env.importDummyCompiler();
+            helper.command.tagScope('1.0.0');
+            helper.command.export(`${forkScope} --include-dependencies --force --set-current-scope --codemod`);
+          });
+          it('should change the files locally on the workspace', () => {
+            const barFoo = helper.fs.readFile('bar/foo.js');
+            expect(barFoo).to.equal(fixtures.barFooModulePath(helper.scopes.remote));
+          });
+          it('should change the dist files locally on the workspace', () => {
+            const barFoo = helper.fs.readFile('bar/foo.js');
+            expect(barFoo).to.equal(fixtures.barFooModulePath(helper.scopes.remote));
+          });
+          it('should change the objects locally', () => {
+            const barFoo = helper.command.catComponent(`${helper.scopes.remote}/bar/foo@latest`);
+            const fileHash = barFoo.files[0].file;
+            const fileContent = helper.command.catObject(fileHash);
+            expect(fileContent).to.have.string(helper.scopes.remote);
+            expect(fileContent).to.not.have.string(forkScope);
           });
         });
       });